requirements.txt
# Browser automation
playwright>=1.40.0

# Testing framework
pytest>=7.4.0
pytest-playwright>=0.4.3

# Reporting
allure-pytest>=2.13.2

# Desktop automation for PingID
pywinauto>=0.6.8
pyperclip>=1.8.2
